+++ apps/kyb-app/src/common/providers/SettingsProvider/settings-provider.context.ts
@@ -1,4 +1,0 @@
-import { ISettings } from '@/common/types/settings';
-import { createContext } from 'react';
-
-export const settingsProviderContext = createContext({} as ISettings);