+++ packages/workflow-core/src/lib/utils/deep-merge/consts.ts
@@ -1,0 +1,6 @@
+export const ARRAY_MERGE_OPTION = {
+  BY_ID: 'by_id',
+  BY_INDEX: 'by_index',
+  CONCAT: 'concat',
+  REPLACE: 'replace',
+} as const;