+++ packages/workflow-core/src/lib/utils/deep-merge/index.ts
@@ -1,0 +1,5 @@
+export * from './consts';
+export * from './deep-merge-with-mutation';
+export * from './deep-merge-with-options';
+export * from './types';
+export * from './utils';