+++ services/workflows-service/src/collection-flow/collection-flow.service.ts
@@ -1,4 +1,8 @@
 import { BusinessService } from '@/business/business.service';
+import {
+  UIElement,
+  UpdateConfigurationDto,
+} from '@/collection-flow/dto/update-configuration-input.dto';
 import { UpdateFlowDto } from '@/collection-flow/dto/update-flow-input.dto';
 import { FlowConfigurationModel } from '@/collection-flow/models/flow-configuration.model';
 import { UiDefDefinition, UiSchemaStep } from '@/collection-flow/models/flow-step.model';
@@ -19,7 +23,7 @@ import { WorkflowDefinitionRepository } from '@/workflow-defintion/workflow-defi
 import { WorkflowRuntimeDataRepository } from '@/workflow/workflow-runtime-data.repository';
 import { WorkflowService } from '@/workflow/workflow.service';
 import { AnyRecord } from '@ballerine/common';
-import { BUILT_IN_EVENT } from '@ballerine/workflow-core';
+import { BUILT_IN_EVENT, deepMergeWithMutation } from '@ballerine/workflow-core';
 import { Injectable } from '@nestjs/common';
 import { EndUser, Prisma, UiDefinition, WorkflowRuntimeData } from '@prisma/client';
 import { randomUUID } from 'crypto';
@@ -87,7 +91,7 @@ export class CollectionFlowService {
       projectIds,
     );
 
-    const uiDefintion = await this.uiDefinitionService.getByWorkflowDefinitionId(
+    const uiDefinition = await this.uiDefinitionService.getByWorkflowDefinitionId(
       workflowDefinition.id,
       'collection_flow' as const,
       projectIds,
@@ -95,7 +99,7 @@ export class CollectionFlowService {
     );
 
     const translationService = new TranslationService(
-      this.getTranslationServiceResources(uiDefintion),
+      this.getTranslationServiceResources(uiDefinition),
     );
 
     await translationService.init();
@@ -103,23 +107,176 @@ export class CollectionFlowService {
     return {
       id: workflowDefinition.id,
       config: workflowDefinition.config,
-      uiOptions: uiDefintion.uiOptions,
+      uiOptions: uiDefinition.uiOptions,
       uiSchema: {
         // @ts-expect-error - error from Prisma types fix
         elements: this.traverseUiSchema(
           // @ts-expect-error - error from Prisma types fix
-          uiDefintion.uiSchema.elements,
+          uiDefinition.uiSchema.elements,
           context,
           language,
           translationService,
         ) as UiSchemaStep[],
+        theme: uiDefinition.theme,
       },
-      definition: uiDefintion.definition
-        ? (uiDefintion.definition as unknown as UiDefDefinition)
+      definition: uiDefinition.definition
+        ? (uiDefinition.definition as unknown as UiDefDefinition)
         : undefined,
     };
   }
 
+  async updateUIDefinition(
+    definitionId: string,
+    payload: UpdateConfigurationDto,
+    projectIds: TProjectIds,
+  ) {
+    const definition = await this.uiDefinitionService.getById(definitionId, {}, projectIds);
+
+    if (!definition) throw new NotFoundException('UI Definition not found');
+
+    this.replaceUIDefinitionElements(definition, payload);
+
+    return await this.uiDefinitionService.updateById(definitionId, {
+      data: {
+        //@ts-ignore
+        uiSchema: definition.uiSchema,
+        //@ts-ignore
+        theme: definition.theme,
+      },
+    });
+  }
+
+  async patchUIDefinition(
+    definitionId: string,
+    payload: UpdateConfigurationDto,
+    projectIds: TProjectIds,
+  ) {
+    const definition = await this.uiDefinitionService.getById(definitionId, {}, projectIds);
+
+    if (!definition) throw new NotFoundException('UI Definition not found');
+
+    this.patchUIDefinitionElements(definition, payload);
+
+    return this.uiDefinitionService.updateById(definitionId, {
+      data: {
+        //@ts-ignore
+        uiSchema: definition.uiSchema,
+        //@ts-ignore
+        theme: definition.theme ?? undefined,
+      },
+    });
+  }
+
+  async deleteUIDefinitionElements(
+    definitionId: string,
+    payload: UpdateConfigurationDto,
+    projectIds: TProjectIds,
+  ) {
+    const definition = await this.uiDefinitionService.getById(definitionId, {}, projectIds);
+
+    if (!definition) throw new NotFoundException('UI Definition not found');
+
+    const { elements = [] } = payload;
+
+    elements.forEach(element => {
+      this.deleteElementByName(
+        element.name,
+        (definition.uiSchema as unknown as { elements: UIElement[] })?.elements as UIElement[],
+      );
+    });
+
+    return this.uiDefinitionService.updateById(definitionId, {
+      data: {
+        //@ts-ignore
+        uiSchema: definition.uiSchema,
+      },
+    });
+  }
+
+  private patchUIDefinitionElements(uiDefinition: UiDefinition, dto: UpdateConfigurationDto) {
+    const { elements = [], theme } = dto;
+
+    elements.forEach(element => {
+      const originElement = this.findElementByName(
+        element.name,
+        (uiDefinition.uiSchema as unknown as { elements: UIElement[] })?.elements as UIElement[],
+      );
+
+      if (originElement) {
+        //@ts-ignore
+        deepMergeWithMutation(originElement, element, {
+          mergeKey: 'name',
+          arrayMergeOption: 'by_index',
+        });
+      }
+    });
+
+    if (theme) {
+      //@ts-ignore
+      uiDefinition.theme = deepMergeWithMutation(uiDefinition.theme, theme);
+    }
+
+    return uiDefinition;
+  }
+
+  private replaceUIDefinitionElements(uiDefinition: UiDefinition, dto: UpdateConfigurationDto) {
+    const { elements = [], theme } = dto;
+
+    elements.forEach(element => {
+      const originElement = this.findElementByName(
+        element.name,
+        (uiDefinition.uiSchema as unknown as { elements: UIElement[] })?.elements as UIElement[],
+      );
+
+      if (originElement) {
+        //@ts-ignore
+        deepMergeWithMutation(originElement, element, {
+          arrayMergeOption: 'replace',
+        });
+      }
+    });
+
+    //@ts-ignore
+    uiDefinition.theme = theme;
+
+    return uiDefinition;
+  }
+
+  private findElementByName(name: string, elements: UIElement[]): UIElement | null {
+    for (const element of elements) {
+      if (element.name === name) {
+        return element;
+      }
+
+      if (element.elements) {
+        const foundInChildren = this.findElementByName(name, element.elements as UIElement[]);
+
+        if (foundInChildren) {
+          return foundInChildren;
+        }
+      }
+    }
+
+    return null;
+  }
+
+  private deleteElementByName(name: string, elements: UIElement[]): UIElement[] {
+    for (const element of elements) {
+      if (element.name === name) {
+        const index = elements.indexOf(element);
+
+        if (index !== -1) {
+          elements.splice(index, 1);
+        }
+      }
+
+      if (element.elements) {
+        this.deleteElementByName(name, element.elements as UIElement[]);
+      }
+    }
+
+    return elements;
+  }
   private getTranslationServiceResources(
     uiDefinition: UiDefinition & { locales?: unknown },
   ): ITranslationServiceResource[] | undefined {