+++ services/workflows-service/src/collection-flow/dto/update-configuration-input.dto.ts
@@ -1,10 +1,21 @@
-import { UiSchemaStep } from '@/collection-flow/models/flow-step.model';
-import { Type } from 'class-transformer';
-import { IsArray, ValidateNested } from 'class-validator';
+import { IsArray, IsObject, IsOptional, IsString } from 'class-validator';
+
+export const CONFIGURATION_UPDATE_STRATEGY = ['patch', 'delete', 'put'] as const;
+export type TConfigurationUpdateStrategy = (typeof CONFIGURATION_UPDATE_STRATEGY)[number];
+
+export class UIElement {
+  @IsString()
+  name!: string;
+
+  elements?: UIElement[];
+}
 
 export class UpdateConfigurationDto {
+  @IsOptional()
   @IsArray()
-  @ValidateNested({ each: true })
-  @Type(() => UiSchemaStep)
-  steps!: UiSchemaStep[];
+  elements?: UIElement[];
+
+  @IsObject()
+  @IsOptional()
+  theme?: object;
 }