+++ services/workflows-service/src/ui-definition/ui-definition.service.ts
@@ -1,7 +1,6 @@
 import type { TProjectId, TProjectIds } from '@/types';
 import { UiDefinitionRepository } from '@/ui-definition/ui-definition.repository';
 import { WorkflowRuntimeDataRepository } from '@/workflow/workflow-runtime-data.repository';
-import { replaceNullsWithUndefined } from '@ballerine/common';
 import { Injectable } from '@nestjs/common';
 import { Prisma, UiDefinitionContext } from '@prisma/client';
 
@@ -81,6 +80,13 @@ export class UiDefinitionService {
       },
       projectIds,
     );
+  }
+
+  async updateById<T extends Omit<Prisma.UiDefinitionUpdateArgs, 'where'>>(
+    id: string,
+    args: Prisma.SelectSubset<T, Omit<Prisma.UiDefinitionUpdateArgs, 'where'>>,
+  ) {
+    return await this.repository.updateById(id, args);
   }
 
   async cloneUIDefinitionById(id: string, projectId: TProjectId) {