+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -243,6 +243,8 @@
 
 ### Patch Changes
 
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.15
 - @ballerine/workflow-core@0.6.15
 
 ## 0.6.14